+++ custom_components/ja_rs485/button.py
@@ -15,7 +15,14 @@ from homeassistant.helpers.dispatcher import async_dispatcher_connect
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
 from .client import JaRs485Client
-from .const import DOMAIN, can_control_pg, is_impulse_pg, is_pg_allowed, signal_update
+from .const import (
+    DOMAIN,
+    can_control_pg,
+    get_custom_name,
+    is_impulse_pg,
+    is_pg_allowed,
+    signal_update,
+)
 from .entity import JaRs485Entity
 
 
@@ -52,7 +59,9 @@ class JaPgButton(JaRs485Entity, ButtonEntity):
         super().__init__(client, entry)
         self.pg_id = pg_id
         self._options = entry.options
-        self._attr_name = f"Jablotron PG {pg_id}"
+        self._attr_name = (
+            get_custom_name(entry.options, "pgs", pg_id) or f"Jablotron PG {pg_id}"
+        )
         self._attr_unique_id = f"{entry.entry_id}_pg_button_{pg_id}"
 
     async def async_press(self) -> None: